Move undo/redo to JS interface
The user can now view a list of past actions and delete individual ones
instead of just undo/redo.

Undo/redo always works with the last item of the history. If you delete
intermediate items, the undo button still removes the last item in the
history, instead of undoing the deletion.