Separate Resolution into Resolution and ResolutionInner
ResolutionInner contains the data of the actual resolution, while
Resolution contains the bells and whistles such as self maps and
callbacks. ResolutionInner is what ResolutionHomomorphism needs to take
in, and is always an immutable object, so is wrapped in Rc<> instead of
Rc<RefCell>.

This separation should make multithreading easier because we only need
ResolutionInner to be Send + Sync. In particular, we don't need the
functions to be Send + Sync.