Eliminate ZeroModule/ZeroModuleHomomorphism
They are replaced with traits of the same name, which says the
Module/ModuleHomomorphism type comes with a method to construct the zero
module/homomorphism. This avoids the need of OptionModule, and in
general makes chain complexes much more pleasant to work with.

These traits were previously included under the names ZeroModuleT and
ZeroModuleHomomorphismT.