Compute Steenrod squares
This has to be the least pretty part of the code ever written. It works
exactly for Steenrod squares at p = 2 for the Adem basis and no more.
Adding support for the Milnor basis (which has an even simpler
comultiplication) should be straightforward, and I will be doing it
soon. Ditto for coproducts at p > 2.

Computing Steenrod squares for higher p requires taking higher tensor
powers. Writing code to tensor two modules was unpleasant enough. People
who care a lot about p > 2 are free to implement this themselves. I do
not see myself doing it.

There is probably also room for optimization. If this code is correct,
all Steenrod operations do not vanish on n apart from Sq^5 (where the
the target group is zero).