Speed up tensor_module by ~50%
A lot of the time was spent computing the operation of the algebra on
FDModules before tensoring them up. Morally, this should be free because
the actions are stored in a table. However, it is not because it is
interfaced via act_on_basis, which requires allocating a new FpVector
and then copying into it.

This commit adds an option to Module that says the result of acting on a
basis element can be borrowed as an &FpVector (this is not possible if
the action has to be calculated, e.g. in FreeModule).