Concurrent resolving
The number of threads used by the resolver in ext-websocket is currently
hardcoded in managers.rs (but not the CLI interface). When using this,
keep in mind that the sseq manager and the browser also require
computing power, so the number of threads used by the resolver should
be less than the number of cores.