Add basic save/load functionality
It currently correctly saves and loads a resolution. The current
implementation has a few restrictions:

 - It doesn't work correctly with sliced vectors/matrices.
 - It doesn't save quasi-inverses and products/self maps.
 - The final binary is uncompressed.

These can be easily fixed if desired. Think of this commit as a proof of
concept.

I am hesitant to produce a fully-implemented save/load functionality, as
it will make it difficult to change formats in the future. For example,
if we want to support sliced vectors, we will have to change how we
store vectors, which would break backwards compatibility.