Add AugmentedMatrix class
This streamlines the step_resolution code. If constant generics were a
thing, we should be able to use this to define AugmentedMatrixN for
every N. If not, we can try to approximate this with macros, but I have
not.

There is some room for improvement, e.g. ways to borrow a segment of a
row. This will hopefully come with a later commit.