Parse, not Validate prime
This introduces the ValidPrime type, which is a wrapper around u32 that is
guaranteed to be a valid prime. This provides a mild speed up and makes
the prime-two implementation cleaner, where now 2 is the only valid
prime.

ValidPrime dereferences to u32, so we can access the underlying value
as *p.