ResolutionHomomorphism::from_module_homomorphism
This lets us produce a ResolutionHomomorphism from a ModuleHomomorphism
between the modules resolved. This only works for chain complexes
concentrated in degree zero.

This further introduces an IdentityHomomorphism trait and implements it
for FDModule and FPModule, which is used in some testing code where we
want to extend the identity and check that we get the identity.

All together, this avoids the need to manual extend_step with
extra_images in many situations. In the long run, we should rethink how
we want to define resolution homomorphisms in general. It is important
to note that the main use case, which is to compute products, does not
fall into the framework above.