Separate libraries and binaries into different workspaces
The reason for this is that every crate in a workspace must be compiled
with the same profile. However, what we want to do is to compile the
library with -O3 while the binary in debug mode (this gives a slowdown
of around 40%, which is much better than full debug builds). This is
possible with sufficiently modern Rust since we can compile dependencies
with a different opt-level than your crate.

There might be better ways to handle this, but this is what I managed.