Allow more flexible module action description
This allows for cumulatively specifying the actions. This is useful for
specifying quotients of the Steenrod algebra in the Milnor basis, since
the multiplication algorithm gives a list of (R, S, T) where P(T) is a
term in P(R) P(S).

(Technically, this commit isn't needed, because each P(T) only shows up
once in the multiplication algorithm. However, this clarifies the intent
of the code)