Add experimental pyodide packager
This is a simpler version of
https://github.com/emscripten-core/emscripten/blob/master/tools/file_packager.py
which pyodide uses to bundle python packages. This only uses the python
standard library, and is tailored for the spectralsequence_chart
package. Specifically, it assumes the library files only consist of .py
files and does not utilize subdirectories.

This does not compress the data using LZ4. The main motivation is to
avoid LZ4 as an external dependency. That said, spectralsequence_chart
is reasonably small to be stored in memory. Further, assuming our
webserver and client support gzip compression, this actually leads to
fewer bytes transferred.

This script appears to produce a reasonable output, but its output has
not been tested.