Pin syn due to breaking changes
There is a breaking change in syn 0.1.58, which breaks enum_dispatch.
enum_dispatch 0.3.5 resolves the issue and is compatible with up-to-date
syn. However, enum_dispatch itself has a regression in 0.3.3:
https://gitlab.com/antonok/enum_dispatch/-/issues/35

So we need to wait for enum_dispatch to fix the regression, and then
update enum_dispatch, which then fixes the syn problem.