ext: Add algorithm to compute Adams d_2's
This uses the secondary Steenrod algebra via the algorithm of
Baues--Jibladze in [1], using the presentation of the secondary
Steenrod algebra of @cnassau in [2].

The code is neither pretty nor extremely efficient, but it appears to
produce correct answers. The results have been verified with the
Isaksen's charts up to (t, s) = (58, 7) and computations up to (t, s) =
(30, 7) have been added as a unit test.

The algorithm is not yet integrated with the rest of the ecosystem but
test_compute_differentials in ext/src/secondary.rs is some logic that
computes and prints the d_2s for the sphere (or rather, the dual of the
d_2's).

For reference, with a release build, it computes d_2(D_1) in 10s on my
computer (one should really read this as "computes all d_2's up to that
of D_1 in 10s"; I'm not aware of a way to compute a single d_2 without
computing all of the rest).

[1] https://arxiv.org/abs/math/0407045
[2] https://arxiv.org/abs/1011.2471