Further optimize PPartMultiplier for p = 2
Manually inline the multinomial function to compute the binomial
coefficient. This saves us from building the vector of diagonal, whcih
is slow. This gives a >15% runtime improvement for secondary.

This slows it down for p > 3 since it is harder to inline the naive way
to inline multinomial is less effective than our algorithm, so we retain
the original version.