Experimental FpVector rewrite
The main problem this is intended to solve is to let us take slices
without having a mutable reference. This should clean up a lot of
existing code. Ideally, this would also lead to cleaner code that is
just as performant. A likely consequence is that we would no longer have
to import the trait FpVectorT to use FpVector, but this is only because
enum_dispatch will not support the new format and we would have to do it
by hand.

The plan is to use const generics to define a FpVectorP type, which is
na FpVector for a fixed P. This allows for monomorphization for any
fixed P without code duplication and type meddling. We then use an enum
to manually wrap the type for various possible values of P and do
enum_dispatch like wrappers. This would also allow us to not hide
internal API's since we are not using a trait.

An FpVectorP is always unsliced. We then introduce a new FpVectorSliceP
type, which represents a slice of FpVectorP, and do the same
dispatching. Most arithmetic functions will be on FpVectorSliceP.