FpVector rewrite (#14)
The main problem this is intended to solve is to let us take slices without having a mutable reference. This should clean up a lot of existing code.

This uses const generics to define a FpVectorP type, which is an FpVector for a fixed prime P. This allows for monomorphization for any fixed P without code duplication and type meddling. We then use an enum
to manually wrap the type for various possible values of P and do enum_dispatch-like wrappers. This would also allow us to not hide internal API's since we are not using a trait.

An FpVectorP is always unsliced. We then introduce a new SliceP type, which represents a slice of FpVectorP, and do the same dispatching. We can implement arithmetic functions separately for `FpVectorP` and `SliceP`, where the former takes advantage of the lack of slicing (i.e. the equivalent of the previous `_pure` versions)