Move resolve_through_bidegree to ResolutionInner
Start migrating to using ResolutionInner instead of Resolution.

This change requires using scoped threads from crossbeam-utils, but we
were already importing that via crossbeam-channels anyway, so we have
not increased the number of dependencies. By virtue of using scoped
threads, we are able to use fewer Arc's around the place.