Allow resolving to non-square region
Previously, if we ran resolve_through_bidegree(5, 10) and then
resolve_through_bidegree(10, 5), we would resolve all the way to (10,
10).

This often incurs extra "hidden" costs when using saved resolutions,
because we may be forced to resolve to much further than we need.

Now it would resolve to the appropriate L-shaped region instead.