Speed up resolution algorithm.
In the previous algorithm, we have to row reduce, add new vectors, then
row reduce again. It is inefficient to run the full row reduction
algorithm again in the second iteration, as most of the matrix is
already row reduced. Further, we already know exactly where the new
pivot columns are. We can use this to produce a more efficient row
reduction algorithm that avoids a lot of the searching.

There is still room for improvement in the last permutation step, but
this already offers a performance improvement.

This delivers a 25% performance improvement in resolving S_2 up to t =
70 using the Adem basis.