New row reduction algorithm
Since RREF's are unique, we can compute them row by row.

At each row, if the row is non-zero, then it is a pivot row for the
first non-zero column. Use this to clear all other rows in this column
and go to the next row.

We then reorder the rows at the end, which is an O(n) operation.

We expect this to be faster because finding the first non-zero column in
a row is faster than finding the first non-zero row in a column. In
practice this delivers a ~10% performance improvement.