Implement M4RI row reduction
When resolving to the 70th stem, this gives a 25% speed up for the Adem
basis and 31% for the Milnor basis. There are some constants that may be
fine-tuned but this is already pretty good.

The M4RI algorithm's efficiency is mostly independent of the density,
which means that the Adem and Milnor basis now have mostly equal
performance.