Don't use RwLock in resolution_wrapper
Just own the things. This simplifies the code a lot and avoids some
scary hacks too.

Fixed a few bugs/logical errors along the way. In particular, we never
need to extend the unit resolution when we extend the main resolution,
and we were doing this in the past for some bizarre reason.