resolve_through_bidegree -> compute_through_bidgree
Use the ChainComplex method for this purpose. To achieve this,
ResolutionHomomoprhism stops calling compute_through_bidegree on the
source/target, and instead just checks that it has been computed with
has_computed_bidegree (otherwise we run into deadlocks).

In essentially all cases we are not relying on this behaviour. Indeed,
computing through bidegree tends to be so much more expensive and
suddenly having to do this expensive operation when calling extend may
be a bit unexpected.

As part of the change, we implemented has_computed_bidegree for all
chain complexes and clarified the contractual guarantees in the case of
resolving to a stem.