Revamp module construction API
ext::utils::construct is now much easier to use, now that we can specify
a module with a string, e.g. "S_2[3]@milnor" means "S_2 shifted by 3,
resolved with the milnor algebra". Both the shift and algebra are
optional arguments. They can also be specified in other forms.

This changes the existing command line format. Instead of supplying the
algebra with --algebra we use the above module format. This means we no
longer need to use an argument parser; simply read the first argument.

In the future I plan to make it possible to supply all arguments either
as command line options or interactively with query, so that commands
can be run in batch. This change should make that easier.

Of course, this change also adds the ability to resolve a shifted
version of a module, which can be useful in multiple situations.