Refactor rust_webserver web interface
The biggest change is the switch to svg-based rendering instead of
canvas-based rendering, using the svg-chart library to handle the
pan/zoom aspects. This allows us to delegate a lot of work to existing
DOM apis. In particular, we no longer have custom code for

 - highlight on hover (use css selectors instead)
 - tooltips (use <title> on the nodes)
 - click handling (use click events on nodes)

There is also a large amount of cleaning up done alongside the changes.

Performance-wise, updates tend to be slower, but I would expect pan/zoom
to be faster.