Refactor secondary Steenrod algebra algorithm.
This is the first step where I strip out all the old code and put back
in a bare bones versin of the new one. This uses a completely different
design that is necessary for the Mod_{Cτ²} algorithm.

Currently, this is still missing concurrency and save file support, but
I'm merging this in because future changes will touch many parts of the
code and it is better to avoid conflict. The old version should not be
used anyway because the old save files discard important data needed for
Mod_{Cτ²}.