Save format 2.0
This is a first draft that implements the ideas in
https://github.com/SpectralSequences/sseq/issues/44#issuecomment-978732350.

Currently, compression is not supported. I think the right way to do
this is that we don't compress when writing, but be able to read
compressed files.

The current implementation loads save files lazily. When we call
resolve_through_stem, for each bidegree, we check if we have data
available. If it is available, we simply load it. Otherwise, we compute
it from scratch as usual.

The save format is designed with the possibility of just mmap'ing the
FpVector limbs. I think mmap'ing a whole matrix is too much complication
as we have do deal with the indirection. rkyv can help you do this but I
don't think it is worth the effort.

A brief documentation of the file format is in ext/SAVE-FORMAT.md

One thing I noticed is that in general the filesystem has overhead for
having a file at all, even if it is tiny. For low degrees this save
format is actually highly inefficient.

Fixes #44