Save format 2.0 (#52)
Redesign resolution save format

The new save format and rationales are documented in `ext/SAVE-FORMAT.md`. We
mention some highlights here.

An immediate beneift of the change is that we can use much less memory when
doing computations. When computing the resolution, once we have computed the
quasi-inverse, we write it to disk, and then we can immediately drop it from
memory. When we have to *use* the quasi-inverse, we can only load the ones we
need and then drop them afterwards.

There is a "streaming" API that prevents the need to load the whole
quasi-inverse; we can load it row by row. This makes it easier to do
distributed computing, where we can use less powerful nodes to do calculations
using a pre-computed resolution.

Another useful change is that each save file has a checksum to mitigate data
corruption, and we also record the algebra used.

The plan is to remove all the saveload code, and use bespoke encoding methods
for each use case. It is currently still used by the secondary Steenrod algebra
code, but it is about to be almost completely rewritten.

Fixes #44