Refactor secondary code again
This is based on the previous code, but rewritten in a way that is more
amenable to parellelization. In particular, we split the
"compute_homotopy" step into two further steps, the first of which is
fully parallelizable with no data dependencies. These two steps can be
made parallel with rayon.