More flexible chain homotopies
When computing a chain homotopy of a chain map F -> G, the chain
homotopy that hits the lowest cohomological degree of G can always be
chosen to be zero when F and G are minimal and G resolves k. Indeed, in
the equation we have to satisfy, changing the bottom-most homotopy only
contributes positive internal degree terms, and the first differential
of G is surjective in positive degrees.

However, a general chain homotopy should allow for the case where the
bottom map is non-zero. This is necessary when trying to compute Massey
products with non-spheres, but also for computing secondary massey
products.