Add prototype for Nassau's algorithm (#31).
The core algorithm works, but there is still a long way to go. The
biggest omission is an algorithm to pick an appropriate subalgebra.
Currently, we only use the largest of F_2, A(0) and A(1) that is
applicable.

There is also much room for optimization. In particular, we are
computing the value of differentials more times than we ought to be.
These values are used when computing the algebra and also when lifting,
so it might be worth saving these on disk. If we do this, we will
probably end up spending more disk space than we used to, but the time
improvement is significant.

With that said, the current algorithm is already faster than the
previous one when resolving up to the 80th stem single-threaded.

We are also missing the following "peripheral" features:

 - Lifting along differentials (implementing `apply_quasi_inverse`)
 - Saving
 - Concurrency support