Support saving Nassau
This works but we should experiment more before considering this final.
Based on some preliminary experiments, the save files are around 20%
larger.

On the other hand, (surprisingly) the quasi-inverses are now faster, at
least for one of the tests I performed. When computing h_0 up to (n, s)
= (80, 40), using Nassau's resolution is around 13% faster. I suspect
this is due to the resolution being less dense instead of a superior
quasi-inverse algorithm.

Most scripts of interest should now work with Nassau's algorithm,
provided a save directory is supplied.