More module trait cleanup
This removes the act_on_basis_borrow function, which was used to
improvem performance of tensor. The implementation when
borrow_output is false is quite bad, so I'm not sure how much
improvement this actually gives compared to a good implementation. In
any case, this can be investigated further when someone revisits the
Steenrod operation code.