secondary: Support more general modules, inc. tmf
Previously, we required the module to be bounded, and checked that there
are no obstructions to lifting all homotopies once and for all. The new
strategy is to check that we lifted correctly whenever we encounter a
potentially problematic lift.

This makes the algorithm more general, since it now works for
unobstructed unbounded modules (such as tmf), and also potentially
obstructed modules where the obstruction in fact vanishes. It also
serves as additional checks for secondary_product and secondary_massey.