Migrate everything to dialog. Fixes #93
To summarize the recent series of commits, all user interactions that
used to use prompt/confrim are now replaced with html dialogs. This
allows for improved ergonomics. For example, we use the placeholder and
pattern attributes for inputs to ensure the inputs are valid before
submission. For class names, we also preview the rendered LaTeX before
the user submits.

The dialog itself is a web component, and currently uses the html dialog
element. This opens up the possibility to migrate to more flexible
dialog implementations (e.g. one that allows moving).

This series also introduces various other reusable web components in
web_ext/sseq_gui/interface/components.js , such as a katex input that
previews the rendered LaTeX when it is not in focus.