Add Quickstart section to README (#195)
I wrote this section for the README after having conversations with several
people who asked whether sseq had a certain feature or how to get started on a
basic level. The goal is to make this more accessible to homotopy theorists who
mostly want to make an Ext chart for their favorite module.

I talked to Joey about this, and also had someone successfully install sseq on
windows using a version of these instructions, but who ran into the "gotchas" I
listed below step 5 in the installation section.